componetiza permissionamento
